This is the final implementation of our project.
There are a few aspects of the project spec that my partner and I chose to leave out, as clarifications regarding certain requirements were made in a late enough stage of the process that a significant amount of our implementation would have to be revised (for example, to allow for multiple instances of an active user, we would need to restructure our implementation of a user to store more of its data in the persistent server, Datastore, and we would have to change a significant amount of code in nearly every function in order to support this). There are also a few opportunities to replace repetitive blocks of code with helper functions, but as this became obvious to us, we were far enough into the project and had documented it well enough that we deemed this unnecessary. However, our final implementation still passed a strong majority of the hidden tests and certainly served its purpose in reinforcing our knowledge of cryptography as well as the language, Go.